requirements.txt
# Парсинг и HTTP-запросы
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.18.1
webdriver-manager==4.0.1

# Работа с данными
pandas==2.2.1
numpy==1.26.4

# База данных
sqlalchemy==2.0.28
aiosqlite==0.19.0

# Веб-сервер
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6

# Визуализация и отчёты
plotly==5.19.0

# Утилиты
python-dotenv==1.0.1
schedule==1.2.1
aiohttp==3.9.3
